Update for new docker version (.raw disk images)
Note that this does not yet support the new directory structure. expects the
images to be in .../com.docker.driver.amd64-linux.

Also, the images must be heterogeneous -- you can't mix .raw and .qcow2
images. Once the base image (Docker.xxx) is changed to a .raw file, any older
.qcow.{name} images will no longer be available or work. This is a limitation
of docker for mac.